Crate nue_codegen

nue derive syntax extension.

Provides the #[derive(PodPacked, Pod, NueEncode, NueDecode)] extensions documented in nue-macros.

Stable

See the syntex documentation for instructions on how to set up your project to use these macros in stable Rust.

Nightly / Unstable

See the example in nue-macros for usage as a normal syntax extension.

Functions

register
Registers the plugin for expansion with syntex.